Move around top-level fns; make EXCLUDE input private
